requirements.txt
aiosqlite
pandas
xgboost
scikit-learn
yfinance
transformers
streamlit>=1.32.0
joblib
huggingface_hub
python-dotenv
torch
lxml
sqlalchemy
psycopg2-binary